Implement Widget trait for MonitorApp and move UI code to ui.rs
Implemented the Widget trait for &mut MonitorApp following ratatui
best practices. All rendering logic has been moved from app.rs to ui.rs.

Benefits:
- MonitorApp is now a proper ratatui widget
- Clean separation: app.rs has state and event loop, ui.rs has rendering
- Easier to test and maintain UI rendering in isolation
- Follows the pattern shown in ratatui examples

The Widget implementation uses helper functions in ui.rs to render
each section (status bar, product, content, hints). All rendering
now operates on Buffer instead of Frame, which is the ratatui way.
